code review comment
Great homepage! I really like the section where you showcase your blog posts on the index page, but it might be clearer if you take out the JS code and place it as an individual file in JS folder. Other than that, everything is perfect! The code is well-organized, and the README is super clear and descriptive. Good job!!!